Added gamepad control to player. Changed the order of some methods at the player class. Added a basic state based behaviour for the player where the state determines the current action the player is taking. The state variable can later be substituted by a enum type of structure.

+++ core/src/com/pinducas/cliche/actors/Actor.java
@@ -14,7 +14,7 @@ public class Actor {
 
 	//DISPOSABLE
 	
-	protected void movimenta(int x, int y){
+	protected void movimenta(float x, float y){
 		body.setTransform(getX() + x, getY() + y, 0);
 	}
